Create README.md
The Appendicular Skeleton

The appendicular skeleton forms the skeleton of the upper and lower limbs. The bones of the upper limb include the pectoral (shoulder) girdle composed of two bones: the clavicle (or collarbone) and the scapula (shoulder blade). The arm bone is called the humerus; the forearm bones are the ulna and the radius. There are eight carpals in each wrist, five metacarpals in each palm, three phalanges in each finger, and two phalanges in each thumb.

The bones of the lower limb are attached to the pelvic girdle, which consists of two hip bones (coxae). Each lower limb contains the thigh bone called the femur, the patella (kneecap), and two leg bones, the tibia (shin bone) and fibula. There are seven tarsals in the ankle, five metatarsals in the middle of each foot, two phalanges in each big toe, and three phalanges in each other toe.